+++ packages/Webkul/Product/src/Exceptions/InsufficientProductInventoryException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Webkul\Product\Exceptions;
+
+use Exception;
+
+class InsufficientProductInventoryException extends Exception {}